Circumvent rectangle draw bug for IE 11
This PR circumvents an IE-11 specific bug in Leaflet Draw whereby it
wasn't possible to draw a rectangle by clicking start and stop points.
To prevent it occuring, we disable that functionality and instead direct
users to click, hold the mouse button down, and drag to draw a
rectangle by updating the tooltip text shown when the rectangle draw
tool is active.

Although the bug didn't affect Chrome, to simplify things we also
disable two-click rectangle drawing there. Polygon drawing wasn't
impacted by the bug and is unaffected here.